requirements.txt
# streamlit==1.32.0
# tensorflow==2.19.0
# numpy==1.26.4
# Pillow

#~~~~~~~~~

gradio==4.31.5
tensorflow==2.15.0
numpy
pillow